swagger add

+++ models/Tickets.go
@@ -1,12 +1,14 @@
 package models
 
+// swagger:model
 type Ticket struct {
-	ID         int    `json:"id"`
+	ID         int64  `json:"id"`
 	Name       string `gorm:"not null" json:"name"`
 	Desc       string `gorm:"not null" json:"desc"`
 	Allocation int    `gorm:"not null" json:"allocation"`
 }
 
+// swagger:model
 type Ticket_Purchase struct {
 	Quantity int    `json:"quantity"`
 	UserID   string `json:"user_id"`
